security/pam_ocra: Fix SSL handling especially wrt FreeBSD 9
The SSL flags were moved to show unconditionally as they are always
correct.  The WITH_OPENSSL_PORT=yes variable was replaced with an
SSL version check so that the port is marked broken if built on
FreeBSD 9 with the base OpenSSL libary selected.  This fixes other
values of SSL_DEFAULT on all platforms.

Approved by:	SSL blanket


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@421788 35697150-7ecd-e111-bb59-0022644237b5