New port: security/rubygem-rex-nop
This library contains the opty2 library for dynamic generation of x86 multi-byte
NOPs. This is useful in writing exploits and encoders. It allows you to dynamic
generate variable length instruction sets that are equivalent to a No Operation
(NOP) without using the actual 0x90 bytecode. The original code was written by
Optyx and spoonm.

WWW: https://github.com/rapid7/rex-nop


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@422314 35697150-7ecd-e111-bb59-0022644237b5