emulators/citra: add new port
Citra is an experimental open-source Nintendo 3DS emulator/debugger
written in C++. It is written with portability in mind, with builds
actively maintained for Windows, Linux and OS X. Citra only emulates a
subset of 3DS hardware, and therefore is generally only useful for
running/debugging homebrew applications. At this time, Citra is even
able to boot several commercial games! Most of these do not run to a
playable state, but we are working every day to advance the project
forward.

https://citra-emu.org


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@423700 35697150-7ecd-e111-bb59-0022644237b5