Unblock the package build via static libressl.  Adapted from [1].
OpenNTPD requires libtls, which is only available from LibreSSL.
If LibreSSL is not the default SSL library, link with a static copy
of the LibreSSL libraries.  This approach has the drawback that
whenever there is an update to libressl, the openntpd port will
require a PORTREVISION bump, otherwise it will not pick up the fixed
library and continue to use the old code.

Drop the RESSL option as LibreSSL is now always available.

PR:		213691 [1]
Submitted by:	jbeich


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@424884 35697150-7ecd-e111-bb59-0022644237b5