ports-mgmt/pkg_search: Remove deprecation, but IGNORE on DragonFly
At the request of a user, further review reveal that the pkg_* tools
aren't actually used.  It just awks the INDEX file.  The maintainer
forgot the technical details, especially since the SEE ALSO section of
the man page references pkg_* tools.

A minor update is needed to re-support DragonFly and update the man
page.  Until this, it's been marked IGNORE for DragonFly.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@425206 35697150-7ecd-e111-bb59-0022644237b5