Simplify Makefile of comms/jsdr
Instead of doing a manual do-configure / do-build / do-install targets, copy
a small 'jsdr.pro' file to WRKSRC and then rely on USES=qmake to do the rest.

The version-suffix of the two binaries spectrum-viewer and fmreceiver was
dropped in the process (why was it there?) -- and therefore the PORTREVISION
is bumped.

PR:		214642
Approved by:	rakuco (mentor)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@426360 35697150-7ecd-e111-bb59-0022644237b5