Import some LXQt core components (required by forthcoming update of x11/qterminal).
LXQt is the Qt port of and the upcoming version of LXDE, the Lightweight
Desktop Environment. It is the product of a merge between the LXDE-Qt and
Razor-qt projects.

This work is based on the LXQt testing repo [1]. Importing is not complete
due to missing pieces in KDE Frameworks (r426118).

I would like to thank contributor and tester, Jesper Schmitz Mouridsen.

[1] https://subversion.assembla.com/svn/lxqt/tags/0.11

Reviewed by:	antoine, rakuco
Obtained from:	LXQt repository
FC after:
Differential Revision:	https://reviews.freebsd.org/D5002


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@426375 35697150-7ecd-e111-bb59-0022644237b5