Update security/gpgme-* to 1.8.0
In this version, libgpgme-pthread.so has been removed in favor of just
using libgpgme.so as the thread-safe library. PORTREVISION has been
bumped on all ports depending on security/gpgme so that any that may have
linked to -lgpgme-pthread will link to -lgpgme instead.

The Python module provided by security/py-gpgme has been renamed upstream
from pyme3 to gpg. This removes the conflict with security/py-pyme,
although security/py-gpgme is still the direct replacement of that
module.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@427185 35697150-7ecd-e111-bb59-0022644237b5