Update to the GCC 6.3 release which brings a number of bugfixes.
Omit the definition of DISTVERSION which is only necessary for ports
tracking snapshots of GCC and simplify the definition of GCC_VERSION
(but keep the latter as a variable to align the gcc6-devel and gcc6 ports).

Remove files/patch-unwind-ia64.h since we have not been supporting ia64
with this release series (ONLY_FOR_ARCHS has not been including ia64).


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@429461 35697150-7ecd-e111-bb59-0022644237b5