[NEW PORT] devel/sonar-ant-task: SonarQube Plugin for analysis projects with an Apache Ant
The Ant SonarQube Task allows integration of SonarQube analysis into an Apache
Ant build script. The Ant Task is a thin wrapper around the SonarQube Runner,
which works by invoking the runner and passing to it all properties named
following a sonar.* convention. This has the downside of not being very Ant-y,
but the upside of providing instant availability of any new analysis parameter
introduced by a new version of a plugin or of SonarQube itself.

PR:	 203075
Submitted by:freebsd dussan.org
Approved by:	adamw (mentor)
Differential Revision:	https://reviews.freebsd.org/D9193


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@431612 35697150-7ecd-e111-bb59-0022644237b5