Fix ocaml-lablgtk2 after r431473 (ocaml-findlib update to 1.7.1).
After this update ocamlfindlib returns two paths separated by a
colon for the destdir, but lablgtk2 configure scripts is not
separating them so it installs things in the wrong place where they
are not grabbed when building the package. Make it extract only the
firsst coponent of the path list.

This fixes build of unison and other dependent ports.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@431617 35697150-7ecd-e111-bb59-0022644237b5