databases/mysql56-server: Do not install my.cnf sample
An issue was discovered where users of mysql did not have a my.cnf and
the recent update was causing mysqld to find a sample my.cnf and load
its parameters. This was causing errors on startup for users of innodb
as the parameters used to init the database did not match the ones in
the sample config file it was now reading.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@431975 35697150-7ecd-e111-bb59-0022644237b5