Import Wayland
This port was first created by kwm@ then updated and improved by Johannes
Lundberg

Wayland is intended as a simpler replacement for X, easier to develop and
maintain. GNOME and KDE are expected to be ported to it.

Wayland is a protocol for a compositor to talk to its clients as well as
a C library implementation of that protocol. The compositor can be a
standalone display server running on Linux kernel modesetting and evdev
input devices, an X application, or a wayland client itself. The clients can
be traditional applications, X servers (rootless or fullscreen) or other
display servers.

Please report bugs to the FreeBSD bugtracker!

WWW: http://wayland.freedesktop.org/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@432406 35697150-7ecd-e111-bb59-0022644237b5