x11/lightdm: Improve rc script to use daemon(8)
This fixes an issue noticed when sysutils/bsdstats was installed and
enabled at runtime in /etc/rc.conf, but for whatever reason was failing
to send the updated statistics. Upon failing, the lightdm process would
receive a SIGHUP, causing it to stop seconds after the login screen was
displayed (before the user even had a chance to login).


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@433952 35697150-7ecd-e111-bb59-0022644237b5