- New port: devel/dub
DUB emerged as a more general replacement for vibe.d's package manager. It does
not imply a dependency to vibe.d for packages and was extended to not only
directly build projects, but also to generate project files (currently VisualD).

Mono-D also supports the use of dub.json (dub's package description) as the
project file.

The project's philosophy is to keep things as simple as possible. All that is
needed to make a project a dub package is to write a short dub.json file and
put the source code into a source subfolder. It can then be registered on the
public package registry to be made available for everyone. Any dependencies
specified in dub.json are automatically downloaded and made available to the
project during the build process.

WWW: https://github.com/dlang/dub


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@435074 35697150-7ecd-e111-bb59-0022644237b5