New port: audio/supercollider
SuperCollider is a programming language for real time audio synthesis
and algorithmic composition.

The language interpreter runs in a cross platform IDE and communicates
via Open Sound Control with one or more synthesis servers.  The
SuperCollider synthesis server runs in a separate process or even on a
separate machine so it is ideal for realtime networked music.

SuperCollider was developed by James McCartney and originally released
in 1996.  He released it under the terms of the GNU General Public
License in 2002 when he joined the Apple Core Audio team.  It is now
maintained and developed by an active and enthusiastic community.  It
is used by musicians, scientists, and artists working with sound.

WWW: http://supercollider.github.io/

PR:		208443
Submitted by:	Tobias Brodel <brittlehaus@gmail.com>
Approved by:	mat (mentor)
Differential Revision:	https://reviews.freebsd.org/D10043


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@436552 35697150-7ecd-e111-bb59-0022644237b5