r436859 removed a bunch of refclock drivers that were enabled by default. (New patch from ian@.)
While here ian@ and I discussed the www/libwww libmd5 dependency. It
turns out that if the sntp configure script finds libmd5, it uses it.
This is now an option to include the dependency.

While working on the libmd5 dependency it was discovered that the
sntp build requires SSL only if it finds libmd5. This is now an
IMPLIES.

Thanks to ian@ for the additional patch and for discovering the
libmd5 check, pointing me in the right direction to solve the without
SSL breakage.

PR:		218078
Submitted by:	ian@ (default driver options)
Discussed with:	ian@ (libmd5 dependency)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@436887 35697150-7ecd-e111-bb59-0022644237b5