Unbreak kipi-plugin-* after r436971.
The kipi-plugin-* ports are all built from the same tarball, which is also used
to build kipi-plugins-kde4 as well.

Upstream does not expect this to happen, and refers to libkipiplugins.so as the
"kipiplugins" target in CMake. Since we build everything separately, each
plugin's build system code does not know this target, which results in
-lkipiplugins (without -L/path/to) being passed to the linker instead of
/path/to/libkipiplugins.so.

Fix it by looking for libkipiplugins.so via CMake and using that result in each
port.

kipi-plugin-ipodexport needs an additional but similar fix, in that it should
look for libgpod using CMake instead of using pkg-config's results directly, as
the latter do not contain full paths.

Reviewed by:	tcberner


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@437106 35697150-7ecd-e111-bb59-0022644237b5