Use native make(1) mechanism to create the package repository directory
In order to have subpackages we need in the end do-package to depend on
a target per package that would be created per subpackages

For this we need to decouple the content of do-package so it becomes
a target that only creates the packages.

As a first step use natural make(1) mechanism for a target: if checks first
for the existence of a file and it it does not exists that execute the content
of the target, in that case it creates the PKGREPOSITORY
Bonus it simplifies a bit the code.

While here to avoid testing multiple time for the PACKAGES directory set a
variable after we tested it the first time

PR:		216877
Approved by:	portmgr (mat)
Reviewed by:	mat
exp-run by:	antoine
Differential Revision:	https://reviews.freebsd.org/D9466


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@438058 35697150-7ecd-e111-bb59-0022644237b5