Deprecate multimedia/banshee
The project cannot cope with a recent (2013) dbus-sharp.  Probably time to seek
for another media player.

Since the port does not build, mark it BROKEN, not DEPRECATED in case someone
wants to jump in and provide a patch, what I am sure not having time nor
interest for in the next weeks :-(.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@440841 35697150-7ecd-e111-bb59-0022644237b5