Update to 3.12.1
Update cad/openvsp to 3.12.1 _and_ unbreak cad/openvsp. The recent
update to eigen3 broke cad/openvsp (since it is used directly) and
graphics/code-eli that is a dependency of cad/openvsp.

Until there is a fix for code-eli change the port to use the internal
eigen3 and code-eli packages bundled in the distribution.

In addition, one of the local patches was accepted upstream and is no
longer necessary.

From the announcement:

Features:
 - Parasite drag buildup tool
 - Full control of VSPAERO from OpenVSP GUI
 - DegenGeom preview visualization
 - Advanced control surface modeling

Bug Fixes:
 - Advanced links not loading from file with 'Loop' policy geoms
 - Crash conformal components with some file airfoils
 - CompGeom when subsurfaces fall on tessellation line
 - DegenGeom handling of tip caps with blended wings
 - CustomGeom name matching GUI mismatch

PR:		220212
Submitted by:	fernando.apesteguia@gmail.com (maintainer)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@444165 35697150-7ecd-e111-bb59-0022644237b5