lang/ruby2{2,3,4}: require libunwind
Since ruby detects this if it's installed, and it's more often installed now
due to other deps, and ruby provides no way to not depend on it if it's found,
pull it in as a dependency unconditionally. While here, fix plist for ruby 2.4
with the CAPIDOCS option on, and restore the MAKE_JOBS_UNSAFE flag when using
that build option since it's not fixed like I thought it was.

PR:		219796
Reported by:	Grzegorz Junka <list1@gjunka.com>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@445549 35697150-7ecd-e111-bb59-0022644237b5