Remove CloudABI Binutils tools.
As of FreeBSD 11, the cloudabi-toolchain port uses the ELF-toolchain
utilities for all the stuff not provided by LLVM. For FreeBSD 10 and
lower, we still rely on Binutils.

As CloudABI itself is only supported as of FreeBSD 11, there is hardly
any interest in supporting a Binutils based toolchain. Let's kick out
the CloudABI Binutils ports and symlink to the utilities in the base
system unconditionally.

Reviewed by:	rene
Differential Revision:	https://reviews.freebsd.org/D11733


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@446997 35697150-7ecd-e111-bb59-0022644237b5