Add PostgreSQL-10 to the ports tree
The PostgreSQL Global Development Group announces today that the
third beta release of PostgreSQL 10 is available for download. This
release contains previews of all of the features which will be
available in the final release of version 10, including fixes to many
of the issues found in the second beta.  Users are encouraged to begin
testing their applications against 10 beta3.

URL:	https://www.postgresql.org/about/news/1771/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@447680 35697150-7ecd-e111-bb59-0022644237b5