Update devel/git to 2.14.1
This version fixes a security issue:

A "ssh://..." URL can result in a "ssh" command line with a
hostname that begins with a dash "-", which would cause the "ssh"
command to instead (mis)treat it as an option.  This is now
prevented by forbidding such a hostname (which will not be
necessary in the real world).

MFH:		2017Q3
Sponsored by:	Rubicon Communications, LLC (Netgate)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@447721 35697150-7ecd-e111-bb59-0022644237b5