net/freeradius3: Disable OpenSSL version checking
FreeRadius developers include a feature enabled by default which checks
your OpenSSL version and refuses to run if certain CVEs are detected.
This is an interesting idea but it means it's possible to upgrade
FreeRadius on a production server and suddently it won't run, especially
if FreeBSD's base OpenSSL doesn't report a version number that can
convince the software it is free from the specified CVEs.

Currently FreeRadius refuses to run on FreeBSD 10.3-RELEASE because it
thinks base system OpenSSL is not patched for CVE-2016-6304, but that
was in fact patched by FreeBSD-10.3-RELEASE-p9.

This feature is only useful if you are using vanilla upstream versions
of OpenSSL which we are not.

Approved by:	portmgr (with hat)
MFH:		2017Q3


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@447967 35697150-7ecd-e111-bb59-0022644237b5