Replace USE_GCC=5+ by USE_GCC=yes in the case of USES=compiler:c++14-lang.
When this was added by r392084 on 2015-07-14, the default flavor of GCC
was GCC 4.8 and explicitly requesting GCC 5 (or later) was necessary for
C++ 14 support.  Now that the default version of GCC is GCC 6, after GCC 5
for several months, we can use the preferred notion of USE_GCC=yes instead
of specifying a concrete minimum version.

Among others this helps with cases where GCC 6 is better adjusted for
FreeBSD, notably the well known std::to_string issue (where that is only
enabled with GCC 6 or later).

PR:		222268
Approved by:	portmgr (antoine)
MFH:		2017Q3


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@449716 35697150-7ecd-e111-bb59-0022644237b5