security/py-{acme,certbot}: update to 0.18.1
Common:

- Update PORTVERSION and distinfo checksum (0.18.1)

security/py-certbot:

- Fix permissions error when upgrading certbot-auto
- Fix user-agent string and help/error message output when the new certbot-auto
  path is used.

If certbot-auto was running as an unprivileged user and it upgraded from
0.17.0 to 0.18.0, it would crash with a permissions error and would need to
be run again to sucessfully complete the upgrade. This has been fixed and
certbot-auto should upgrade cleanly to 0.18.1.

Certbot usually uses "certbot-auto" or "letsencrypt-auto" in error-messages
and the User-Agent string instead of "certbot" when you are using one of
these wrapper scripts. Proper detection of this was broken with Certbot's new
installation path in /opt in 0.18.0 but this problem has been resolved.

More details about these changes can be found on the GitHub repo:

https://github.com/certbot/certbot/milestone/45?closed=1

Changelog:

https://github.com/certbot/certbot/blob/master/CHANGELOG.md#0180---2017-09-06

Reviewed by:	koobs (maintainer)
Approved by:	koobs (maintainer, py-certbot)
MFH:		2017Q3
Differential Revision:	https://reviews.freebsd.org/D12362


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@449834 35697150-7ecd-e111-bb59-0022644237b5