finance/py-stripe: Revert (part of) r426059
Revert r426059 which removed py-unittest2.

(1) The port was revived and (2) the change was incomplete, breaking the
test target/phase because tests were not confirmed to continue to work
(without downloading the dependency, which is not allowed in poudriere),
or updated to work without unittest2.

Removing *_DEPENDS alone for whatever reason (expiry/deletion in this case)
is insufficient for ports of Python packages, as their code may still
register, reference or attempt to import that (package) code, and fatally
error if it is not available.

Other ports modified in that revision very likely require similar
updates, if they haven't been updated or fixed already.

[1] svnweb.freebsd.org/changeset/ports/426059


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@453089 35697150-7ecd-e111-bb59-0022644237b5