New port: audio/cava
C.A.V.A. is a bar spectrum audio visualizer for the terminal using
ALSA, PulseAudio, sndio or fifo buffer for input.

This program is not intended for scientific use.  It's written to look
responsive and aesthetic when used to visualize music.

WWW: https://karlstav.github.io/cava

- Include a backport of an sndio input contributed by me

PR:		223301
Submitted by:	omar.polo@protonmail.com (based on)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@453161 35697150-7ecd-e111-bb59-0022644237b5