New port: multimedia/kodi-devel
Kodi is a free and open source cross-platform media-player and
entertainment hub.  It allows users to play and view most videos,
music, podcasts, and other digital media files from local and network
storage media and the internet.

This is a development snapshot of Kodi for the upcoming Kodi v18.

Special thanks to ahtotaat for working with upstream to get Kodi to
build and work out of the box on FreeBSD.

Changes:	http://kodi.wiki/view/Kodi_v18_(Leia)_changelog


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@456215 35697150-7ecd-e111-bb59-0022644237b5