New port: chinese/ibus-cangjie
This is an IBus engine for users of the Cangjie and
Quick input methods.

It is primarily intended to Hong Kong people
who want to input Traditional Chinese,
as they are (by far) the majority of Cangjie and Quick users.

WWW: http://cangjians.github.io/projects/ibus-cangjie/

PR:		223565
Submitted by:	hello@blubee.me


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@457716 35697150-7ecd-e111-bb59-0022644237b5