Stop truncating .MAKEOVERRIDES after the FLAVORS work.
There is no such variable as ${MAKEOVERRIDES}, only ${.MAKEOVERRIDES}, so this
code was simply truncating .MAKEOVERRIDES entirely which could cause performance
and other odd issues with OSREL and make recursion.

We don't want to leak FLAVOR to child sub-makes but that can be fixed later;
this code was not doing what it intended.

Approved by:	portmgr (implicit)
MFH:		2018Q1


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@458048 35697150-7ecd-e111-bb59-0022644237b5