Fix flavor helpers when only bsd.port.pre.mk is not used and FLAVOR is not set.
The helpers were being evaluated before the default FLAVOR was set. The
helpers worked fine if the FLAVOR was specified by the user but not if
it was left to the default value. Resolve this by moving the helpers
right after the default FLAVOR is set.

PR:		224879
exp-run by:	antoine
Approved by:	portmgr (implicit)
MFH:		2018Q1
Differential Revision:	https://reviews.freebsd.org/D13756


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@458394 35697150-7ecd-e111-bb59-0022644237b5