net-im/py-pyxmpp: mark deprecated
This port is outdated (last release was made in 2011), unmaintained and has no
consumers in ports tree. According to the official github page "this project is
becoming obsolete" since 2011 in favor of PyXMPP2 (from the same author, not yet
in ports) that support modern Python API's and run with both Python branches.

Mark it deprecated and set to expire at 2018-02-28.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@460427 35697150-7ecd-e111-bb59-0022644237b5