Fix build when gettext is installed
When building, if gettext(1) is found, CMake generates rules to update the
cpp-hocon.pot translation file.  However, the generated rules for ninja(1) are
broken and break the build.

Always rely on regular make(1) in order to avoid this situation.

Not bumping PORTREVISION since successfully built packages are not affected by
this change.

Reported by:	marius.schuller@makandra.de (by mail to puppet@)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@461788 35697150-7ecd-e111-bb59-0022644237b5