FLAVORS: Reorder so the default is first.
It is documented that the first FLAVOR is the default.  Due to some use
of DEFAULT_VERSIONS and USES, the default FLAVOR may be determined
dynamically.  Poudriere in particular expects the default to be first
and does the wrong thing for emacs ports currently [1].  This resolves
that and avoids other tools making the same mistake.

PR:		225659 comment #14
Approved by:	portmgr (implicit)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@464273 35697150-7ecd-e111-bb59-0022644237b5