Update dovecot to 2.3.1, and dovecot-pigeonhole to 0.5.1
This is a very large update, and it WILL require manually
updating existing conf files, though the changes to do so
are not extensive. Updating instructions are here:

	https://wiki2.dovecot.org/Upgrading/2.3

Additionally there are various cleanups to the dovecot rc(8)
script, and support for a LUA scripting interface for dovecot.

The decision was made not to import the 2.3.0 or 2.3.0.1 releases
here, due to the number of existing bugs. ler and I have been
dogfooding it for months now, and all of the bugs I've encountered
are fixed in this 2.3.1 release.

This update is the result of many, many hours of collborative work
between ler and me, and the input of many people on the freebsd-ports
list.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@466172 35697150-7ecd-e111-bb59-0022644237b5