Add benchmarks/uperf
Unified Performance Tool (or uperf for short) is a network
performance tool that supports modelling and replay of various
networking patterns. It represents a new kind of benchmarking tool
(like filebench) where instead of running a fixed benchmark or
workload, a description (or model) of the workload is provided and
the tool generates the load according to the model.

PR:		226649
Submitted by:	0mp@FreeBSD.org
Requested by:	gnn


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@466231 35697150-7ecd-e111-bb59-0022644237b5