Fix astro/qmapshack ahead of CMake 3.11.
Because of the way CMake 3.11 orders includes in mocs_compilation.cpp,
it shows up cases where headers don't have all of the necessary
includes or forward-type-declarations in place. Add one here to
fix the build. The error message is in the patch.

Also reported upstream,
  https://bitbucket.org/maproom/qmapshack/issues/325/cmake-311-build-fix
but it had already been (unrelatedly) fixed upstream in
  https://bitbucket.org/maproom/qmapshack/commits/2968b238

Approved by:	tcberner (mentor, implicit)
Differential Revision:	https://reviews.freebsd.org/D14932


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@466261 35697150-7ecd-e111-bb59-0022644237b5