Add authors file for informative author fields with `git log`
From the top of the ports tree run
	git config svn.authorsfile .gitauthors
and `git log` will show author fields that look like
	Rodney Grimes <rgrimes@FreeBSD.org>
instead of
	user <user@aed309b6-a8cd-e111-996c-001c23d10e55>.

Warning: If a commit is made by an author who is not included in this
file, `git svn dcommit` will fail until the author has been added.

This file is append-only.  Anyone who has ever committed or has commit
access to any of the repositories should be included.  The only exception
when users could be removed is if they no longer have an account and
never committed.

TODO: - Add peter's script [1] to a post-commit hook, so that additions to
        svnadmin/conf/access will be reflected in .gitauthors as soon as
        possible.
      - Add peter's script to a cron job on freefall to update .gitauthors
        when users change their gecos information.

[1] Latest version of the script can be found at
    freefall:/home/jrm/update_gitauthors.pl

Reviewed by:	eadler karels mat peter rgrimes swills
Differential Revision:	https://reviews.freebsd.org/D13754


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@467739 35697150-7ecd-e111-bb59-0022644237b5