New port: devel/samurai
samurai is a ninja-compatible build tool written in C99 with a focus on
simplicity, speed, and portability.

It implements the ninja build language through version 1.8.2 except
for MSVC dependency handling (deps = msvc).  It uses the same format
for .ninja_log and .ninja_deps as ninja, currently version 5 and 3
respectively.

It is largely feature-complete and supports most of the same options as
ninja.

WWW: https://github.com/michaelforney/samurai


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@468415 35697150-7ecd-e111-bb59-0022644237b5