New port: devel/p5-Call-Context
Sanity-check calling context

If your function only expects to return a list, then a call in some
other context is, by definition, an error. The problem is that,
depending on how the function is written, it may actually do something
expected in testing, but then in production act differently.

WWW: http://search.cpan.org/dist/Call-Context/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@469622 35697150-7ecd-e111-bb59-0022644237b5