net-im/mastodon: Mark broken and deprecated and set expiration date
File bankruptcy with net-im/mastodon.  After the www/node upgrade to 10, it
broke again.  In general, the return to effort ratio trying to contort
these node modules and ruby gems in to a port is too low (for me).  I will
leave the port here for about a month in case anyone else would like to
give it a go.

For those who want to run a FreeBSD-based instance on the "fediverse", it
is probably easier to simply follow Mastodon's installation guide.
Alternatively, I have heard good things about Pleroma [1], so you may want
to give it a try.

[1] https://git.pleroma.social/pleroma/pleroma


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@470717 35697150-7ecd-e111-bb59-0022644237b5