Upgrade to Apache OpenOffice trunk SVN revsion r1833124.  This includes the bug fix in extra-patch-align16 needed for amd64. [1]
Instead of symlinking the directory containing the OpenOffice .desktop
files under $PREFIX/share/applications, symlink the individual .desktop
files.  This is what upstream does on Linux and what LibreOffice does
as well.  The Plasma 5 desktop ignores symlinks to directories when it
is scanning for .desktop files.

PR:		228030 [1]
Reported by:	kan


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@472230 35697150-7ecd-e111-bb59-0022644237b5