New port: devel/p5-Test-Deep-UnorderedPairs
Plugin for Test::Deep to compare unordered lists of tuples.

This module provides the sub unordered_pairs (and tuples, samehash,
as synonyms) to indicate the data being tested is a list of pairs
that should be tested where the order of the pairs is insignificant.

WWW: https://metacpan.org/release/Test-Deep-UnorderedPairs

PR:		229195
Submitted by:	Tomohiro Hosaka <bokutin@bokut.in>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@472959 35697150-7ecd-e111-bb59-0022644237b5