Add py-xhtml2pdf 0.2.2
xhtml2pdf is a html2pdf converter using the ReportLab Toolkit, the HTML5lib and
pyPdf. It supports HTML 5 and CSS 2.1 (and some of CSS 3). It is completely
written in pure Python so it is platform independent.

The main benefit of this tool that a user with Web skills like HTML and CSS is
able to generate PDF templates very quickly without learning new technologies.

WWW: https://github.com/xhtml2pdf/xhtml2pdf


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@473216 35697150-7ecd-e111-bb59-0022644237b5