Add py-ratelimiter 1.2.0
This package provides the ratelimiter module, which ensures that an operation
will not be executed more than a given number of times on a given period. This
can prove useful when working with third parties APIs which require for example
a maximum of 10 requests per second.

WWW: https://github.com/RazerM/ratelimiter


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@473217 35697150-7ecd-e111-bb59-0022644237b5