science/mpqc: Fix library dependencies
Problems in this port caused the science/libghemical breakage
(it was labeled as DEPRECATED).

Two major fixes:
* Fix the unknown 'backtrace' symbol: libexecinfo.so dependency was missing for libSCmisc.so
* Fix broken libint.so: -L${LOCALBASE} was missing for libSCcints.so and libint.so was just ignored
  and effectively missing

These two problems were previously attempted to be patched with "-Wl,--as-needed -lexecinfo -Wl,--no-as-needed"
but this didn't work. So this line is now removed from Makefile, and proper flags are supplied to
the link commands.

Other port changes:
* Take maintainership
* Add missing RUN_DEPENDS for wish:x11-toolkits/tk-wrapper
* Add CONFLICTS_BUILD
* Fix MPICH_LIB_DEPENDS to point to net/mpich2
* patch the 'wish' path in configure


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@474828 35697150-7ecd-e111-bb59-0022644237b5