net/mpd5: revert previous LCP MRU negotiation change and add new way to deal with same problem following upstream changes r22478, r2249:
- reverse r2246 as it used wrong way to deal with the problem;
- make it possible to override results of LCP negotiation
  and set interface MTU to specified value with a command:

set iface mtu NNN override

New option "override" is not documented yet.
This is dangerous option as it can violate RFC 1661
and should be used with caution and at your own risk.
It is useful when you deal with broken PPP peer
negotiating too low value while higher MTU is known to work.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@476338 35697150-7ecd-e111-bb59-0022644237b5