graphics/mupdf: Fix linking with libmupdf.so for third-party ports
libmupdf.so references symbols from libmupdfthird.so but does not
link directly against it.

This is problematic for ports using the --as-needed linker option,
such as graphics/zathura-pdf-mupdf (>= 0.3.3) due to Meson's defaults.
zathura-pdf-mupdf itself does not reference symbols from libmupdfthird.so,
and so despite being included in the linker arguments, libmupdfthird.so
is not actually linked in.

$ zathura --version
error: Could not load plugin '/usr/local/lib/zathura/libpdf-mupdf.so' (/usr/local/lib/libmupdf.so.1.13.0: Undefined symbol "js_tointeger").

This is a consequence of us providing shared libraries in spite of
upstream since r433550 to keep package sizes down, but not properly
understanding the dependency between libmupdfthird and libmupdf.

PR:		232059
Submitted by:	Albert Ou <aou@eecs.berkeley.edu>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@481606 35697150-7ecd-e111-bb59-0022644237b5