net/intel-{ix,em}-kmod: freeze PORTREVISION and PORTEPOCH
PORTREVISION wasn't specified in net/intel-{ix,em}-kmod, which are
slave ports of net/intel-ixl-kmod. Because of that when PORTREVISION
was bumped in intel-ixl-kmod - it automatically appeared in the slave
ports, which shouldn't happen.

To avoid this in the future - explicitly specify PORTREVISION and
PORTEPOCH in the slave ports.

Sanity and compile tested using poudriere w/portlint.

Approved By: sbruno (mentor), franco_opnsense.org (intel-em-kmod maintainer)

Differential Revision: https://reviews.freebsd.org/D17646


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@483387 35697150-7ecd-e111-bb59-0022644237b5