audio/csound6: Take maintainership; Change to github
This port is broken. It both failed to fetch and failed to build afterwards.
I fixed the fetching problem: it moved to github.
So far I couldn't fix the build problem - it seems to fail due to some incompatibilities
with dependencies, but it is also badly outdated. New version doesn't build either.

The upstream needs to fix at least these two problems:
* https://github.com/csound/csound/issues/1063
* https://github.com/csound/csound/issues/1064

I'll come back to it after they fix these build issues.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@485176 35697150-7ecd-e111-bb59-0022644237b5