deskutils/kmail: include optional.h from boost to fix build on recent CURRENT
On FreeBSD Current optional is no longer in experimental/optional.
Further, switching  to c++17 to use it, would require changes to
databases/akonadi to no longer use auto_ptr.

As this all would be a hassle, follow what upstream has done for a
similar problem with akonadi on Windows [1]: Import a 3rdparty optional.h
and use that.

[1] https://github.com/KDE/akonadi/commit/482315f4cbc79d3725738fa1e62c8e0ce053bdea

PR:             234022


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@488694 35697150-7ecd-e111-bb59-0022644237b5