lang/php73: Make enabling MYSQL native driver optional
Currently PHP is always compiled with --enable-mysqlnd, to allow the use
of the native MySQL Native Driver. MySQL Native Driver is a replacement for the
MySQL Client Library.

While this is handy when working with MySQL there is no need for
it when MySQL is *not* used at all. This happens frequently when
working without databases or simply with other databases.

To avoid POLA the newly introduced option is a default option.

Disabling it will reduce the size of the package by ca. 175 KB,
which also helps in modern constraint VM run environments.

Submitted by:	Reko Turja <reko.turja@liukuma.net>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@490542 35697150-7ecd-e111-bb59-0022644237b5