New port: x11-fonts/league-gothic
League Gothic is a revival of an old classic, Alternate Gothic #1. It was
originally designed by Morris Fuller Benton for the American Type Founders
Company in 1903. The company went bankrupt in 1993, and since the original
typeface was created before 1923, the typeface is in the public domain.

WWW: https://www.theleagueofmoveabletype.com/league-gothic

PR:		231304
Submitted by:	Florian Limberger <flo@snakeoilproductions.net>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@491129 35697150-7ecd-e111-bb59-0022644237b5