Restore a simpler and more elegant version of r490816 which was reverted in r490825. (r490827 addresses the breakage of INDEX which happened in some scenarios that led to that revert.)
Add a new option, GCC, that controls the use of GCC instead of the
system compiler (usually clang these days) as we have been doing since
r334935 in 2013-11-26.  And make that the default, both since there've
been user reports about Wine not working properly with the toolchain
in FreeBSD 12 and since there may be other challenges given that
upstrea and in other distributions GCC is being used all along.

PR:		231010
Reviewed by:	mat
Tested by:	sbruno


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@491148 35697150-7ecd-e111-bb59-0022644237b5