graphics/drm-current-kmod: Update messges
Update pkg-message and makefile COMMENT to remove references to
drm-devel-kmod and the mention that this is the development version.  It is
the version for FreeBSD CURRENT, so the ride might be a little bumpy, but
it's not the devlopment version.

No changes to package.

FreeBSDDesktop issue:	#129

Reported by:	grahamperrin
Sponsored by:	B3 Init (zeising)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@491609 35697150-7ecd-e111-bb59-0022644237b5